basic upgradable token
