+++ smart_contracts/.gitignore
@@ -5,7 +5,7 @@ coverage.json
 typechain
 typechain-types
 
-# Hardhat files
+#Hardhat files
 cache
 artifacts
