+++ smart_contracts/contracts/Stablecoin.sol
@@ -1,0 +1,37 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.4;
+
+import "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
+import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
+import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
+
+contract Stablecoin is Initializable, ERC20Upgradeable, OwnableUpgradeable, UUPSUpgradeable {
+    /// @custom:oz-upgrades-unsafe-allow constructor
+    constructor() {
+        _disableInitializers();
+    }
+
+    function initialize() initializer public {
+        __ERC20_init("Stablecoin", "stk");
+        __Ownable_init();
+        __UUPSUpgradeable_init();
+
+        _mint(msg.sender, 1000000 * 10 ** decimals());
+    }
+
+    function _authorizeUpgrade(address newImplementation)
+        internal
+        onlyOwner
+        override
+    {}
+
+        function _beforeTokenTransfer(address from, address to, uint256 amount)
+        internal
+        override
+    {
+        super._beforeTokenTransfer(from, to, amount);
+        
+    }
+
+}
